sécurité des communications/Mai 2025/TP1/Partie 4/requirements.txt
rsa
requests
flask
mysql-connector-python
flask_socketio